# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for rev tracker.",
        "title": "RevTracker server Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "erupshis",
            "email": "e.rupshis@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/homework_question/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Select godoc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework_question"
                ],
                "summary": "select homework question",
                "operationId": "hw-question-select",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework question id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.HomeworkQuestion"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update godoc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework_question"
                ],
                "summary": "updates homework question",
                "operationId": "hw-question-update",
                "parameters": [
                    {
                        "description": "updated homework question",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.HomeworkQuestion"
                        }
                    },
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.HomeworkQuestion"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete godoc",
                "tags": [
                    "homework_question"
                ],
                "summary": "deletes existing homework_question",
                "operationId": "hw-question-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/question/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Select godoc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "select question",
                "operationId": "question-select",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Question"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update godoc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "updates question",
                "operationId": "question-update",
                "parameters": [
                    {
                        "description": "updated question",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Question"
                        }
                    },
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Question"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete godoc",
                "tags": [
                    "question"
                ],
                "summary": "deletes existing question",
                "operationId": "question-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.Content": {
            "type": "object",
            "properties": {
                "Answer": {
                    "type": "string"
                },
                "Solution": {
                    "type": "string"
                },
                "Task": {
                    "type": "string"
                }
            }
        },
...
        "data.HomeworkQuestion": {
            "type": "object",
            "properties": {
                "Homework_Id": {
                    "type": "integer"
                },
                "Id": {
                    "type": "integer"
                },
                "Order": {
                    "type": "integer"
                },
                "Question_Id": {
                    "type": "integer"
                }
            }
        },
        "data.Question": {
            "type": "object",
            "properties": {
                "Content": {
                    "$ref": "#/definitions/data.Content"
                },
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}